Moving ObservationEvent and its subclasses to a separate class-file. Adding the concept of TestGoal. Renaming FTA to just TestAgent
